TEST: Se crea una vista abstracta para la creación de los pedidos para el camarero y cliente duplicidad SONAR
